Update build_folder_structure.R
I think I have solved the problem of adding in a rmarkdown template in to the correct folder. I still get a weird error message but it works (small victories!!) The error is 
" Error in editor(file = file, title = title) : Error in parse(targetFile) : attempt to use zero-length variable name
You can attempt to correct the error using \TheDataPackage\Test\minimum_metadata\minimum_metadata\minimum_metadata.Rmd = edit() "
Maybe an issue with the "package name" (in this case Test) is not created until the function is run. Possible solution is to have a separate function to build the templates (build_templates())?